requirements.txt
Django==4.1.7
djangorestframework==3.14.0
django-ckeditor==6.5.1
psycopg2==2.9.5
drfdocs==0.0.11